Separate contextmanager tests
